generate_file_list bug fixes and remove file locking
Fixed an issue in WIN32 version of generate_file_list(), where depending on compiler options the function would sometimes not work.  This was resolved by specifying a non-unicode version of FindFirstFile().  Also, a bug where folders beginning with a dot ('.') were considered to be valid folders, was fixed.  Additionally, file locking had been previously removed for WIN32, and now all calls to file locking for unix are commented out as well. (fps_lock and fps_unlock).  These can be uncommented if their use is desired.