now using _stat64 calls to handle larger data
Replacing fstat() and _stat() with _stat64() calls.  This allows for file_sizes greater than 4 GB.  Under Windows, the old way (fstat and _stat) would yield a value of zero for the file_size in the case where the file size is greater than ~4GB (more than 32 bits to represent).  Non-Windows code is unchanged.