reverting read_MEF_segment() change
In read_MEF_segment() a previous change made it so files within segments are always close after reading, even if only the header has been read.  As this contradicts the published documentation for meflib (which states that files should be left open if only the header is read, "MEF 3 Specification", page 72), this update reverts to the original default behavior.  However, a check for a precompiler definition "FORCE_CLOSE_AFTER_READ" is added for situations where the application may not be able to handle very many open files at a time.