reverting previous change
After conversation - we decided the behavior of FPS_DIRECTIVE_CLOSE_FILE_DEFAULT should not be overwritten by other meflib routines, so with that directive set to MEF_TRUE, then the files should always be closed after the segment is read.  The documentation for read_MEF_segment() will be updated accordingly.