bug fix for extract_path_parts on WIN32
fixed extract_path_parts on WIN32 - before there was a potential crash bug in the case where there was no path (no slashes) and no extension (no dot).  Now it properly pre-pends the current path in the case where no path is specified.  This mirrors the behavior of the non-WIN32 version of the function.  This is non-unicode code, so updates may be necessary where unicode is required.