updated MEF_FULL_FILE_NAME_BYTES
Rather than hardcoding a full file path length to 1024 (255 unicode) characters maximum, use the OS built-in _MAX_PATH (for Windows) and PATH_MAX (for unix).  This number is still very small when using Windows, so adjustments may need to be made to accommodate longer paths.  Such changes may be addressed in future updates.